move the enumerator mess out of the main Sequence.cs; better pinned usage in Reference<T>
